Correct errors in vid_ratings function
e.g., an error due to change from having 3 video rating blocks to only have 2 (as videos are only rated once prior to completing the task in the final version, but were rated twice in earlier versions). Also changed the code so that it extracts ratings of points punishment (as planned).